Split useInnerWatch/illegal-usage.spec
