+++ tests/helpers.ts
@@ -1,4 +1,4 @@
-import { Compare } from "../src"
+import { Compare, InnerStore } from "../src"
 
 export abstract class Counter {
   public abstract readonly count: number
@@ -26,4 +26,32 @@ export abstract class Counter {
   public static inc({ count }: Counter): Counter {
     return { count: count + 1 }
   }
+}
+
+export interface WithStore<T = Counter> {
+  store: InnerStore<T>
+}
+
+export interface WithFirst<T = Counter> {
+  first: InnerStore<T>
+}
+
+export interface WithSecond<T = Counter> {
+  second: InnerStore<T>
+}
+
+export interface WithThird<T = Counter> {
+  third: InnerStore<T>
+}
+
+export interface WithSpy {
+  spy: VoidFunction
+}
+
+export interface WithListener {
+  listener: VoidFunction
+}
+
+export interface WithInitial<T = Counter> {
+  initial?: T
 }
